Move rotate defition to a top-level section.
...with some other stuff in it.

I also decided to change the notation for rotation; the current version
is inspired by FIPS, and has the concrete advantage of having 'L' in the
name to make the direction of the shift clear from the call site.

I tried to also define a bitwise or operator and use that to define
ROTL, but for some reason pandoc doesn't want to render TeX's \or
operator. For now I just caved and continued to use addition, but it
would be nice to define the bitwise or for the 'choice of M' section as
well; it's a bit weird that we're using C there and nowhere else.