Add a pdf target, and check in the generated file
Keeping a copy of the generated spec makes it easier to casually read
on the web: there's no need to clone the repository or install pandoc.
It also means one has to make sure to index the generated file before
each commit, but that can be taken care of by a (client-side) hook if
desired.