Adjust Makefile rule for spec.pdf.
Apparently the version of pandoc pulled in on the 'ubuntu-latest' image
doesn't recognize `-t pdf`; instead it wants you to use `-t latex` and
will infer pdf from the output filename (so we have to use -o). Both
approaches seem to work on my machine, so I guess let's use the lowest
common denominator.

This also changes the html rule to use -o, just for consistency.