Replace m and v parameters with T.
The mask and value scheme is more general, but the leveling algorithm
only makes sense when you're checking for a number of bottom bits, and
afaik no system actually does anything more complex.

This had been originally added to capture the difference between perkeep
and bup -- the former checks for trailing zeros, while the latter checks
for trailing ones. But I'm increasingly of the opinion that it's better
not to over-parametrize this, and just be prescriptive; standardization
is the whole point, after all.